+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0 - 2024-12-13]
+## [1.1.0 - 2024-12-19]
 ### Added
 - Dockerfile
 - Single Spa
@@ -80,6 +80,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Cancel Reject button
 - Archive button requires message too
 - Current User in Project View
+- Press create button only if mandatory fields are not empty
 
 ### Changed
 - Rename accept and reject bridgehead state buttons (authorize/revoke)